requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.4.0
dj-database-url==2.2.0
Django==5.1.1
django-flags==5.0.13
djangorestframework==3.15.2
Faker==30.6.0
gunicorn==23.0.0
idna==3.10
model-bakery==1.20.0
numpy==2.1.3
packaging==24.1
pillow==10.4.0
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
urllib3==2.2.3
whitenoise==6.7.0
